+++ zscaler/zpa/services/extranet_resource/extranet_resource.go
@@ -2,7 +2,9 @@ package extranet_resource
 
 import (
 	"context"
+	"fmt"
 	"net/http"
+	"strings"
 
 	"github.com/zscaler/zscaler-sdk-go/v3/zscaler"
 	"github.com/zscaler/zscaler-sdk-go/v3/zscaler/zpa/services/common"
@@ -15,9 +17,23 @@ const (
 
 func GetExtranetResourcePartner(ctx context.Context, service *zscaler.Service) ([]common.CommonSummary, *http.Response, error) {
 	relativeURL := mgmtConfig + service.Client.GetCustomerID() + extranetResourceEndpoint
-	list, resp, err := common.GetAllPagesGenericWithCustomFilters[common.CommonSummary](ctx, service.Client, relativeURL, common.Filter{MicroTenantID: service.MicroTenantID()})
+	list, resp, err := common.GetAllPagesGenericWithCustomFilters[common.CommonSummary](ctx, service.Client, relativeURL, common.Filter{})
 	if err != nil {
 		return nil, nil, err
 	}
 	return list, resp, nil
+}
+
+func GetExtranetResourcePartnerByName(ctx context.Context, service *zscaler.Service, extranetName string) (*common.CommonSummary, *http.Response, error) {
+	relativeURL := mgmtConfig + service.Client.GetCustomerID() + extranetResourceEndpoint
+	list, resp, err := common.GetAllPagesGenericWithCustomFilters[common.CommonSummary](ctx, service.Client, relativeURL, common.Filter{})
+	if err != nil {
+		return nil, nil, err
+	}
+	for _, app := range list {
+		if strings.EqualFold(app.Name, extranetName) {
+			return &app, resp, nil
+		}
+	}
+	return nil, resp, fmt.Errorf("no extranet resource named '%s' was found", extranetName)
 }